Ajout Patch Endpoint

+++ Escapade/ControllerUser.cs
@@ -14,6 +14,7 @@ using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Enums;
 using Microsoft.OpenApi.Models;
 using System.Net;
 using Swashbuckle.AspNetCore.Annotations;
+using Microsoft.AspNetCore.JsonPatch;
 
 namespace AzureFunctionEscapade
 {
@@ -110,9 +111,7 @@ namespace AzureFunctionEscapade
                 var user = await _userService.GetById(id);
 
                 if (user == null)
-                {
                     return new UnprocessableEntityObjectResult($"No user exists with id: {id}");
-                }
 
                 return new OkObjectResult(user);
             }
@@ -125,20 +124,28 @@ namespace AzureFunctionEscapade
             }
         }
 
-        [FunctionName("UpdateUser")]
-        [OpenApiOperation(operationId: "Run", tags: new[] { "UpdateUser" })]
+        [FunctionName("UpdateUserPost")]
+        [OpenApiOperation(operationId: "Run", tags: new[] { "UpdateUserPost" })]
         [OpenApiSecurity("function_key", SecuritySchemeType.ApiKey, Name = "code", In = OpenApiSecurityLocationType.Query)]
         [OpenApiParameter(name: "id", In = ParameterLocation.Path, Required = true, Type = typeof(string), Description = "L'identifiant de l'utilisateur")]
         [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "text/plain", bodyType: typeof(string), Description = "The OK response")]
-        public async Task<IActionResult> UpdateUser([HttpTrigger(AuthorizationLevel.Anonymous, "put", Route = "users/{id}")] HttpRequest req,
+        [OpenApiRequestBody(
+        "application/json",
+        typeof(User),
+        Description = "The user to update.")]
+        public async Task<IActionResult> UpdateUserPost([HttpTrigger(AuthorizationLevel.Anonymous, "put", Route = "users/{id}")] HttpRequest req,
             ILogger log, string id)
         {
             var userJson = await req.ReadAsStringAsync();
 
             try
             {
-                var user = JsonConvert.DeserializeObject<User>(userJson);
-                user.Id = id;
+                var user = await _userService.GetById(id);
+
+                if (user == null)
+                    return new UnprocessableEntityObjectResult($"No user exists with id: {id}");
+
+                user = JsonConvert.DeserializeObject<User>(userJson);
 
                 await _userService.Update(user);
 
@@ -146,7 +153,47 @@ namespace AzureFunctionEscapade
             }
             catch (Exception e)
             {
-                var errorMessage = $"Failed to update book with id: {id} with details: {userJson}";
+                var errorMessage = $"Failed to update user with id: {id} with details: {userJson}";
+
+                log.LogError(e, errorMessage);
+                return new BadRequestObjectResult(errorMessage);
+            }
+        }
+
+        [FunctionName("UpdateUserPatch")]
+        [OpenApiOperation(operationId: "Run", tags: new[] { "UpdateUserPatch" })]
+        [OpenApiSecurity("function_key", SecuritySchemeType.ApiKey, Name = "code", In = OpenApiSecurityLocationType.Query)]
+        [OpenApiParameter(name: "id", In = ParameterLocation.Path, Required = true, Type = typeof(string), Description = "L'identifiant de l'utilisateur")]
+        [OpenApiResponseWithBody(statusCode: HttpStatusCode.OK, contentType: "text/plain", bodyType: typeof(string), Description = "The OK response")]
+        [OpenApiRequestBody(
+        "application/json",
+        typeof(JsonPatchDocument<User>),
+        Description = "The patch document for updating the user.")]
+        public async Task<IActionResult> UpdateUserPatch(
+        [HttpTrigger(AuthorizationLevel.Anonymous, "patch", Route = "users/{id}")] HttpRequest req,
+        ILogger log, string id)
+        {
+            var patchDocumentJson = await req.ReadAsStringAsync();
+
+            try
+            {
+                var user = await _userService.GetById(id);
+
+                if (user == null)
+                    return new UnprocessableEntityObjectResult($"No user exists with id: {id}");
+
+                var patchDocument = JsonConvert.DeserializeObject<JsonPatchDocument<User>>(patchDocumentJson);
+
+                // Appliquer le document de patch à l'utilisateur existant
+                patchDocument.ApplyTo(user);
+
+                await _userService.Update(user);
+
+                return new OkObjectResult(user);
+            }
+            catch (Exception e)
+            {
+                var errorMessage = $"Failed to update user with id: {id} with patch details: {patchDocumentJson}";
 
                 log.LogError(e, errorMessage);
                 return new BadRequestObjectResult(errorMessage);
@@ -163,6 +210,12 @@ namespace AzureFunctionEscapade
         {
             try
             {
+
+                var user = await _userService.GetById(id);
+
+                if (user == null)
+                    return new UnprocessableEntityObjectResult($"No user exists with id: {id}");
+
                 await _userService.Delete(id);
 
                 return new NoContentResult();
